+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * First pass implementing absorbtion for monsters (needs midi qol).
 * Effects: Support Potent Cantrip for midi-qol.
 * Effects: If using midi-qol Bardic Inspiration effect is now applied to feature.
+* Action parsing: code tidy.
 
 # 2.7.6